+++ crates/services/src/task_manager/dkim.rs
@@ -33,7 +33,7 @@ use store::{
     },
     write::now,
 };
-use trc::DkimEvent;
+use trc::{DkimEvent, DnsEvent};
 use types::id::Id;
 
 pub(crate) trait DkimManagementTask: Sync + Send {
@@ -473,6 +473,20 @@ async fn dkim_management(server: &Server, task: &TaskDomainManagement) -> trc::R
     // Delete signatures
     for signature in delete_signatures {
         let record = generate_dkim_dns_record_name(&signature.object, &domain.name);
+
+        if let Some((updater, origin)) = &dns_updater
+            && let Err(err) = updater
+                .set_rrset(origin, &record, dns_update::DnsRecordType::TXT, Vec::new())
+                .await
+        {
+            trc::event!(
+                Dns(DnsEvent::RecordDeletionFailed),
+                Hostname = record.clone(),
+                Details = origin.clone(),
+                Type = "TXT",
+                Reason = err.to_string(),
+            );
+        }
 
         trc::event!(
             Dkim(DkimEvent::SignatureDeleted),